created pagination apis

+++ src/main/java/edu/neu/cs6510/repositories/LookUpRepository.java
@@ -47,7 +47,7 @@ public interface LookUpRepository extends CrudRepository<LookUpData, Integer> {
             "where attribute_mapping_id in ?1 and (-1 = ?6 or year in ?2) " +
             "and (COALESCE( null, ?3) is null or location_id in ?3) and value between ?4 and ?5) as a "  +
             "join gov_fin_location_info on a.location_id = gov_fin_location_info.id", nativeQuery = true)
-    Integer queryLookUpDataTotal(@Param("attributeId")List<Integer> attributeId, @Param("year") List<Integer> year,
+    Long queryLookUpDataTotal(@Param("attributeId")List<Integer> attributeId, @Param("year") List<Integer> year,
                                  @Param("locationId") List<Integer> locationId,
                                  @Param("from")Integer from, @Param("to")Integer to, @Param("yearSize") Integer yearSize);
 
@@ -63,7 +63,7 @@ public interface LookUpRepository extends CrudRepository<LookUpData, Integer> {
             "where attribute_mapping_id in ?1 and (-1 = ?6 or year in ?2) " +
             "and value between ?4 and ?5 and (COALESCE( null, ?3) is null or location_id in (select id from gov_fin_location_info where type_code = ?3))) as a " +
             "join gov_fin_location_info on a.location_id = gov_fin_location_info.id", nativeQuery = true)
-    Integer queryLookUpDataTotal(@Param("attributeId") List<Integer> attributeId, @Param("year") List<Integer> year,
+    Long queryLookUpDataTotal(@Param("attributeId") List<Integer> attributeId, @Param("year") List<Integer> year,
                                  @Param("typeCode") Integer typeCode, @Param("from") Integer from,
                                  @Param("to")Integer to,@Param("yearSize") Integer yearSize);
 
@@ -81,7 +81,7 @@ public interface LookUpRepository extends CrudRepository<LookUpData, Integer> {
             "where attribute_mapping_id in ?1 and (-1 = ?6 or year in ?2) " +
             "and value between ?4 and ?5 and (COALESCE( null, ?3) is null or location_id in (select id from gov_fin_location_info where parent_id = ?3))) as a " +
             "join gov_fin_location_info on a.location_id = gov_fin_location_info.id", nativeQuery = true)
-    Integer queryLookUpDataParentIdTotal(@Param("attributeId") List<Integer> attributeId, @Param("year") List<Integer> year,
+    Long queryLookUpDataParentIdTotal(@Param("attributeId") List<Integer> attributeId, @Param("year") List<Integer> year,
                                          @Param("parentId") Integer parentId,@Param("from")  Integer from,@Param("to")  Integer to,
                                          @Param("yearSize") Integer yearSize);
